Add context menu and fixes

+++ Lantean.QBTMudBlade/Interop/BoundingClientRect.cs
@@ -2,16 +2,16 @@
 {
     public class BoundingClientRect : ClientSize
     {
-        public int Bottom { get; set; }
+        public double Bottom { get; set; }
 
-        public int Top { get; set; }
+        public double Top { get; set; }
 
-        public int Left { get; set; }
+        public double Left { get; set; }
 
-        public int Right { get; set; }
+        public double Right { get; set; }
 
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
     }
 }
