+++ Lantean.QBTMudBlade/Interop/ClientSize.cs
@@ -2,8 +2,8 @@
 {
     public class ClientSize
     {
-        public int Width { get; set; }
+        public double Width { get; set; }
 
-        public int Height { get; set; }
+        public double Height { get; set; }
     }
 }
